rename all the view file
