+++ INSTRODUCTION.md
@@ -9,23 +9,15 @@ index.js // (业务逻辑代码入口)
 tpl.js // (模板拼装入口)
 模块名.ejs // (页面编写入口)
 
-/**
- * 以下可根据自己需要添加
- * 模块下可以建立文件目录用来填写业务代码,上述只是基础规范，扩展性比较强
- * /
-eg:
-  src/views/test-demo/ 在该目录下创建文件夹/aboutCom
-        ---- a.js 业务a代码
-        ---- b.js 业务b代码
-都是需要通过ES6规范导出导入
+// 可执行 npm run new 自动创建基础页面配置
 ```
 * 页面公用css，全部需要在公用base.js引入(便于webpack处理),[配置](https://github.com/BiYuqi/webpack-seed/blob/master/src/common/js/base.js),然后每个页面引入base.js [使用](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/index/index.js#L4)
 * 页面开发跳转页面都是基于打包后输出的绝对路径进行编写 **html/模块.html** [详情](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/index/index.ejs)
 * 全部采用模块化开发，**每个模块都是一个文件夹** [详情](https://github.com/BiYuqi/webpack-seed/tree/master/src/views) (开发环境views下创建)
-* 该文件夹包含 **模块模板写页面(模块名.ejs) + 模板混合(tpl.js) + index.js（该模块业务逻辑）** 打包后会自动注入，无需手动引入js文件 [详情](https://github.com/BiYuqi/webpack-seed/tree/master/src/views/test-demo)
+* 该文件夹包含 **模块模板写页面(模块名.ejs) + 模板混合(tpl.js) + index.js（该模块业务逻辑）** 打包后会自动注入，无需手动引入js文件 [详情](https://github.com/BiYuqi/webpack-seed/tree/master/src/views/test-page)
 * 各个js功能模块之间互相引用，一律使用ES6语法
-* 样式编写基于各模块入口js直接 **import '样式地址 '** [参考](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-demo/index.js#L5) 
-* 页面(.ejs)--图片引入方式 [详情](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-demo/test-demo.ejs#L10)
+* 样式编写基于各模块入口js直接 **import '样式地址 '** [参考](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-page/index.js#L5) 
+* 页面(.ejs)--图片引入方式 [详情](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-page/test-page.ejs#L10)
 
 * **assets是webpack resolve配置好的别名，代表assets目录**
 ```html
@@ -48,7 +40,7 @@ eg:
 | **src/**
 | ---common/ 项目公用资源 (公用图片,css,js等配置)
 | ----------------libs.js 第三库自动渲染到页面(此处配置的静态资源,会自动注入到页面) [配置](https://github.com/BiYuqi/webpack-seed/blob/master/src/common/libs/libs.js) [注入页面](https://github.com/BiYuqi/webpack-seed/blob/master/src/components/footer/footer.ejs) [底层实现](https://github.com/BiYuqi/webpack-seed/blob/master/src/layout/layout/layout.js#L5)
-| ---api 接口请求配置 [配置](https://github.com/BiYuqi/webpack-seed/blob/master/src/api/request.js) [编写](https://github.com/BiYuqi/webpack-seed/blob/master/src/api/movie.js) [使用](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-demo/index.js#L9)
+| ---api 接口请求配置 [配置](https://github.com/BiYuqi/webpack-seed/blob/master/src/api/request.js) [编写](https://github.com/BiYuqi/webpack-seed/blob/master/src/api/movie.js) [使用](https://github.com/BiYuqi/webpack-seed/blob/master/src/views/test-page/index.js#L9)
 | ---components 项目公用组件部分可以存在此处, 比如自己开发的插件
 | ---layout/ 项目结构模板 (供各个子模块调用，后续可扩展多样化模板,可以添加不带侧边栏的模板等)
 | ----------------base 页面基础模板
