+++ src/views/login-page/index.js
@@ -2,7 +2,7 @@
  * 默认样式+默认逻辑
  */
 import '@/common/js/base'
-import './login.scss'
+import './login-page.scss'
 import __ from 'utils/dom'
 
 __('.wbs-login_btn').on('click', () => {
