+++ src/views/login-page/tpl.js
@@ -1,4 +1,4 @@
-import content from './login.ejs'
+import content from './login-page.ejs'
 import layout from 'layout/layoutAuth/layoutAuth.js'
 const pageTitle = '登录'
