+++ src/views/news/news-page/tpl.js
@@ -1,6 +1,6 @@
-import content from './test-demo.ejs'
+import content from './news-page.ejs'
 import layout from 'layout/layout/layout.js'
-const pageTitle = '测试页面'
+const pageTitle = '新闻-科技'
 
 const temp = layout.init({ pageTitle }).run(content())
