+++ src/views/test-page/__test__/index-test.js
@@ -1,7 +1,7 @@
 import { expect } from 'chai'
 
 describe('What is the good unit test?', () => {
-  it('The good test is to test user behavior', () => {
+  it('should render correct value', () => {
     const res = 'good test'
     expect(res).to.equal('good test')
   })
